moved tests to fourier test file.
